+++ Releases/Control Enemies Mod/Changelog.md
@@ -1,5 +1,6 @@
 ## v1.0.3 (unreleased)
 - Updated to work with new RocketLib version.
+- Fix players being able to be trapped in hell portal in competitive mode if they don't currently have enough points to attempt a win.
 
 ## v1.0.2
 - Updated to work with new RocketLib version.